fixed an egregious spelling error
