+++ chapter_9/exercise_9_1_and_9_2/exercise_9_2.py
@@ -1,12 +1,12 @@
-# exercise 9-2 three resturants
-# make three resturants and call describe resturant on all three instances
+# exercise 9-2 three restaurants
+# make three restaurants and call describe restaurant on all three instances
 # sigh...
-from resturant import Resturant as R
+from restaurant import Restaurant as R
 
 res_1 = R('the grotto', 'new californian cusine')
 res_2 = R('taj mahal', 'indian')
 res_3 = R('grad school', 'greasy spoon')
 
-res_1.describe_resturant()
-res_2.describe_resturant()
-res_3.describe_resturant()
+res_1.describe_restaurant()
+res_2.describe_restaurant()
+res_3.describe_restaurant()
